// src/types.ts
export type Theme = Pick<
  InferSelectModel<typeof themes>,
  | "id"
  | "name"
  | "previewUrl"
  | "colors"
  | "fontFamily"
  | "fontFamilyHeader"
  | "fontSizes"
  | "fontWeights"
  | "imageMaskUrl"
  | "backgroundImageUrl"
  | "isCorporate"
  | "assets"
  | "isPublic"
>;